fix(mapper): attempt to trim distribution version
Unsee supports a range of Alertmanager versions, and maps between the
Alertmanger format and the internal Unsee format using mappers. It
determines what mapper to use by parsing the Alertmanager version on the
assumption it's semver compatible. Some distributions of Alertmanager
append distribution specific metadata to this version string, which the
semver package isn't tolerant of.

This CL creates an internal semver package that sits between Unsee and
the upstream semver package. On calls to `MustParse` it first attempts
to strip the distribution metadata by tossing out everything after the
first tilde.

Bug: #263
Change-Id: I64e66af9ebfe82ae9e55d72f4a049d645a9de299